First step towards modularising app.py Split out Utility functions into api_helpers.py
 Changes to be committed:
	modified:   app.py
	new file:   services/api_helpers.py
